redux folders reorganized, component folders reorganized initially to check
