+++ frontend/src/api/questions.ts
@@ -1,8 +1,8 @@
 import axios from "axios";
-import { URL } from "../../enum";
+import { URL } from "../enum";
 
 const getAll = async () => {
-  const response = await axios.get(URL.EMPLOYEES_URL);
+  const response = await axios.get(URL.QUESTIONS_URL);
   return response.data;
 };
