--- frontend/src/components/hr/analytics/HRAnalytics.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { UserRole } from "../../../enum";
-import { useAppDispatch, useAppSelector } from "../../../redux/hooks/hooks";
-
-const Analytics = () => {
-  const { selectedRole } = useAppSelector((state) => state.loginUser);
-
-  if (selectedRole === UserRole.HR) {
-    return <div>Analytics</div>;
-  } else return null;
-};
-
-export default Analytics;
