+++ frontend/src/components/hr/dashboard/DBUsers.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from "react";
-import { initialiseEmployees } from "../../../redux/reducers/user/userListSlice";
-import { useAppDispatch, useAppSelector } from "../../../redux/hooks/hooks";
+import { useAppDispatch, useAppSelector } from "../../../hooks/hooks";
 import Table from "react-bootstrap/Table";
-import { IEmployee } from "../../../redux/types/userTypes";
-import { RootState } from "../../../redux/store";
+import { IEmployee } from "../../../types/userTypes";
+import { RootState } from "../../../app/store";
+import { initialiseEmployees } from "../../../features/user/userListSlice";
 
 const DBUsers = () => {
   const dispatch = useAppDispatch();
