--- frontend/src/components/hr/fileFolders/FileFolders.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { useAppDispatch, useAppSelector } from "../../../redux/hooks/hooks";
-import { UserRole } from "../../../enum";
-
-const FileFolders = () => {
-  const { selectedRole } = useAppSelector((state) => state.loginUser);
-
-  if (selectedRole === UserRole.HR) {
-    return <div>FileFolders</div>;
-  } else return null;
-};
-
-export default FileFolders;
