--- frontend/src/components/hr/inbox/HRInbox.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-import { useAppDispatch, useAppSelector } from "../../../redux/hooks/hooks";
-import { UserRole } from "../../../enum";
-
-const Inbox = () => {
-  const { selectedRole } = useAppSelector((state) => state.loginUser);
-
-  if (selectedRole === UserRole.HR) {
-    return <div>HR Inbox</div>;
-  } else if (selectedRole === UserRole.User) {
-    return <div>Component for employee role</div>;
-  } else if (selectedRole === UserRole.Manager) {
-    return <div>Component for manager role</div>;
-  } else return null;
-};
-
-export default Inbox;
