+++ frontend/src/components/hr/questionnaire/QuestionMult.tsx
@@ -2,13 +2,10 @@ import React from "react";
 import Form from "react-bootstrap/Form";
 import Accordion from "react-bootstrap/Accordion";
 import { Card } from "react-bootstrap";
-import { RootState } from "../../../redux/store";
-import {
-  useAppSelector,
-  useAppDispatch,
-} from "../../../../src/redux/hooks/hooks";
-import { initialiseQuestions } from "../../../redux/reducers/survey/surveySlice";
-import { IQuestion } from "../../../redux/types/dataTypes";
+import { RootState } from "../../../app/store";
+import { useAppSelector, useAppDispatch } from "../../../hooks/hooks";
+import { initialiseQuestions } from "../../../features/survey/surveySlice";
+import { IQuestion } from "../../../types/dataTypes";
 
 function QuestionMult() {
   const dispatch = useAppDispatch();
