+++ frontend/src/components/login/Login.tsx
@@ -3,9 +3,8 @@ import jwt_decode from "jwt-decode";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
 import { UserRole } from "../../enum";
-import { initialiseEmployees } from "../../redux/reducers/user/userListSlice";
-import { initialiseQuestions } from "../../redux/reducers/survey/surveySlice";
-import { useAppDispatch, useAppSelector } from "../../../src/redux/hooks/hooks";
+
+import { useAppDispatch, useAppSelector } from "../../hooks/hooks";
 import classes from "./Login.module.css";
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
@@ -13,7 +12,9 @@ import {
   setIsAuthenticated,
   setSelectedRole,
   setUserEmail,
-} from "../../redux/reducers/login/loginSlice";
+} from "../../features/login/loginSlice";
+import { initialiseEmployees } from "../../features/user/userListSlice";
+import { initialiseQuestions } from "../../features/survey/surveySlice";
 
 interface LoginProps {}
