+++ frontend/src/components/shared/analytics/Analytics.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { UserRole } from "../../../enum";
-import { useAppDispatch, useAppSelector } from "../../../redux/hooks/hooks";
+import { useAppDispatch, useAppSelector } from "../../../hooks/hooks";
 
 const Analytics = () => {
   const { selectedRole } = useAppSelector((state) => state.loginUser);
