+++ frontend/src/components/shared/dashboard/Dashboard.module.css
@@ -1,0 +1,38 @@
+.dashboard_container {
+  display: grid;
+  grid-template-columns: repeat(25, 1fr);
+  grid-template-rows: repeat(25, 1fr);
+}
+
+.dashboard_chart {
+  background: white;
+  grid-column: 2/17;
+  grid-row: 1/12;
+  border-radius: 8px;
+}
+
+.dashboard_surList {
+  background: white;
+  grid-column: 2/17;
+  grid-row: 13/25;
+  border-radius: 8px;
+  max-height: 500px;
+  overflow-y: auto;
+}
+
+.dashboard_progress {
+  background: white;
+  grid-column: 18/26;
+  grid-row: 1/12;
+  border-radius: 8px;
+}
+
+.dashboard_emplList {
+  background: white;
+  grid-column: 18/26;
+  grid-row: 13/25;
+  max-height: 500px;
+  overflow-y: auto;
+  padding: 10px;
+  overflow-y: scroll;
+}
