+++ frontend/src/components/shared/inbox/Inbox.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useAppDispatch, useAppSelector } from "../../../redux/hooks/hooks";
+import { useAppDispatch, useAppSelector } from "../../../hooks/hooks";
 import { UserRole } from "../../../enum";
 
 const Inbox = () => {
@@ -8,7 +8,7 @@ const Inbox = () => {
   if (selectedRole === UserRole.HR) {
     return <div>HR Inbox</div>;
   } else if (selectedRole === UserRole.User) {
-    return <div>Component for employee role</div>;
+    return <div>Inbox Component for employee role</div>;
   } else if (selectedRole === UserRole.Manager) {
     return <div>Component for manager role</div>;
   } else return null;
