+++ frontend/src/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
-import { store } from "./redux/store";
+import { store } from "./app/store";
 import App from "./App";
 import "./index.css";
 import { BrowserRouter } from "react-router-dom";
