+++ frontend/src/routes/MainRoutes.tsx
@@ -1,103 +1,39 @@
 import React from "react";
 import { Route, Routes, Navigate } from "react-router-dom";
 import { UserRole } from "../enum";
-import { useAppSelector } from "../../src/redux/hooks/hooks";
+import { useAppSelector } from "../hooks/hooks";
 import Info from "../components/userProfile/Info";
 import Logout from "../components/login/Logout";
 import MyProfile from "../components/userProfile/MyProfile";
 import Main from "./templates/main/Main";
-import HRDashboard from "../components/hr/dashboard/HRDashboard";
-import HRInbox from "../components/hr/inbox/HRInbox";
-import HRSurveys from "../components/hr/surveys/HRSurveys";
-import HRUsers from "../components/hr/users/HRUsers";
-import HRQuestionnaire from "../components/hr/questionnaire/HRQuestionnaire";
-import HRAnalytics from "../components/hr/analytics/HRAnalytics";
-import HRFilesFolders from "../components/hr/fileFolders/FileFolders";
-import ManagerDashboard from "../components/manager/dashboard/ManagerDashboard";
-import ManagerInbox from "../components/manager/inbox/ManagerInbox";
-import ManagerSurveys from "../components/manager/surveys/ManagerSurveys";
-import UserDashboard from "../components/user/dashboard/UserDashboard";
-import UserInbox from "../components/user/inbox/UserInbox";
 import CreateSurvey from "../components/hr/surveys/CreateSurvey";
+import Dashboard from "../components/shared/dashboard/Dashboard";
+import Inbox from "../components/shared/inbox/Inbox";
+import Surveys from "../components/shared/surveys/Surveys";
+import Users from "../components/shared/users/Users";
+import Questionnaire from "../components/shared/questionnaire/Questionnaire";
+import Analytics from "../components/shared/analytics/Analytics";
+import FileFolders from "../components/shared/fileFolders/FileFolders";
 
 const MainRoutes = () => {
-  const { isAuthenticated } = useAppSelector((state) => state.loginUser);
   const { selectedRole } = useAppSelector((state) => state.loginUser);
-
-  if (!isAuthenticated) {
-    return <Navigate to="/" />;
-  }
 
   return (
     <div>
       <Routes>
         <Route path="/" element={<Main />}>
-          <Route
-            index
-            element={
-              selectedRole === UserRole.HR ? (
-                <HRDashboard />
-              ) : selectedRole === UserRole.Manager ? (
-                <ManagerDashboard />
-              ) : selectedRole === UserRole.User ? (
-                <UserDashboard />
-              ) : null
-            }
-          ></Route>
-          <Route
-            path="/home"
-            element={
-              selectedRole === UserRole.HR ? (
-                <HRDashboard />
-              ) : selectedRole === UserRole.Manager ? (
-                <ManagerDashboard />
-              ) : selectedRole === UserRole.User ? (
-                <UserDashboard />
-              ) : null
-            }
-          ></Route>
-          <Route
-            path="/inbox"
-            element={
-              selectedRole === UserRole.HR ? (
-                <HRInbox />
-              ) : selectedRole === UserRole.Manager ? (
-                <ManagerInbox />
-              ) : selectedRole === UserRole.User ? (
-                <UserInbox />
-              ) : null
-            }
-          ></Route>
-          <Route
-            path="/surveys"
-            element={
-              selectedRole === UserRole.HR ? (
-                <HRSurveys />
-              ) : selectedRole === UserRole.Manager ? (
-                <ManagerSurveys />
-              ) : null
-            }
-          ></Route>{" "}
+          <Route index element={<Dashboard />}></Route>
+          <Route path="/home" element={<Dashboard />}></Route>
+          <Route path="/inbox" element={<Inbox />}></Route>
+          <Route path="/surveys" element={<Surveys />}></Route>{" "}
           <Route
             path="/createsurvey"
             element={selectedRole === UserRole.HR ? <CreateSurvey /> : null}
           />
-          <Route
-            path="/users"
-            element={selectedRole === UserRole.HR ? <HRUsers /> : null}
-          ></Route>
-          <Route
-            path="/questionnaire"
-            element={selectedRole === UserRole.HR ? <HRQuestionnaire /> : null}
-          ></Route>
-          <Route
-            path="/analytics"
-            element={selectedRole === UserRole.HR ? <HRAnalytics /> : null}
-          ></Route>
-          <Route
-            path="/filesfolders"
-            element={selectedRole === UserRole.HR ? <HRFilesFolders /> : null}
-          ></Route>
+          <Route path="/users" element={<Users />}></Route>
+          <Route path="/questionnaire" element={<Questionnaire />}></Route>
+          <Route path="/analytics" element={<Analytics />}></Route>
+          <Route path="/filesfolders" element={<FileFolders />}></Route>
           <Route path="/myprofile" element={<MyProfile />}></Route>
           <Route path="/info" element={<Info />}></Route>
           <Route path="/logout" element={<Logout />}></Route>
