+++ frontend/src/types/loginTypes.ts
@@ -1,4 +1,4 @@
-import { UserRole } from "../../enum";
+import { UserRole } from "../enum";
 
 export interface ILogin {
   isAuthenticated: boolean;
